Move cli options into separate file
