+++ scripts/schemas/artifact_schema.ts
