+++ scripts/schemas/compiler_schema.ts
@@ -1,0 +1,10 @@
+export interface CompilerOptions {
+    contractsDir: string;
+    networkId: number;
+    optimizerRuns: number;
+    artifactsDir: string;
+}
+
+export interface ContractSources {
+    [key: string]: string;
+}
